# Conflicts:
#	server/src/prepared_statements/assignment_ps.ts
#	server/src/routes/assignments.ts

+++ server/src/routes/assignments.ts
@@ -122,10 +122,12 @@ const updateAssignment = async function(req: any, res: any) {
             // Assemble the file path. Updated file name is the new file name.
             // It can never be the old since req.file would be undefined.
             const newFilePath = path.join(fileFolder, updatedFileName);
-            const oldFilePath = path.join(fileFolder, oldFilename);
 
             // Remove the old file and write the new file.
-            await fs.unlink(oldFilePath);
+            if (oldFilename) {
+                const oldFilePath = path.join(fileFolder, oldFilename);
+                await fs.unlink(oldFilePath);
+            }
             await fs.writeFile(newFilePath, req.file.buffer);
         }
         res.json(result);
@@ -139,16 +141,17 @@ const updateAssignment = async function(req: any, res: any) {
 // Function which adds the assignment to the database.
 const addAssignmentToDatabase = async function(req: any, res: any) {
     try {
-        // Error in case of no file
+        let fileName: string | null;
+        let filePath: string | undefined = undefined;
+
         if (req.file == undefined) {
-            res.status(400);
-            res.json({ error: "No file uploaded" });
-            return;
+            // tslint:disable-next-line
+            fileName = null;
+        } else {
+            fileName = Date.now() + "-" + req.file.originalname;
+            filePath = path.join(fileFolder, fileName);
         }
 
-        const fileName = Date.now() + "-" + req.file.originalname;
-        const filePath = path.join(fileFolder, fileName);
-        // add to database
         const result: any = await AssignmentPS.executeAddAssignment(
             req.body.title,
             req.body.description,
@@ -161,8 +164,11 @@ const addAssignmentToDatabase = async function(req: any, res: any) {
             req.body.review_due_date,
             req.body.one_person_groups,
             req.body.review_evaluation);
-        // writing the file if no error is there
-        await fs.writeFile(filePath, req.file.buffer);
+
+        if (filePath) {
+            // writing the file if no error is there
+            await fs.writeFile(filePath, req.file.buffer);
+        }
 
         // Generate a default review evaluation rubric if review evaluation is turned on.
         if (result.review_evaluation) {
